feat(core): f32 motion-state storage — slice 1 of f32-hot-columns
state_n2, sampled/root/trace poses (Pose32), and capture vectors store
f32; every accessor keeps its f64 signature (round on store, widen on
load) and all arithmetic between stays f64, so the oracle remains a
bit-exact gate. Bail-path substitution rounds draws to match capture
storage; env-captured lexical values round with the capture vector and
overlay the interpreter env (capture_env) so interpreted paths see the
same rounded bindings; the cull-reused-pose oracle expected side
rounds through Pose32. Round-trip and f32 bail-parity tests added.

Gates first-hand: 349 core tests plain + MAKU_LOWER_ORACLE=1, 6
release ignored oracle suites, scripted outcomes unchanged.

+++ crates/core/src/sim/slots.rs
@@ -324,7 +324,7 @@ fn sample_curve_collider_frac(
     projection: &CapsuleChainSlot,
     frame: Pose,
     capture_layout: Option<&CaptureLayout>,
-    captures: &[f64],
+    captures: &[f32],
     tick_rate: f64,
 ) -> Option<Vec<(f64, f64)>> {
     sample_curve_projection(
@@ -350,7 +350,7 @@ fn sample_curve_projection(
     u_max: f64,
     frame: Pose,
     capture_layout: Option<&CaptureLayout>,
-    captures: &[f64],
+    captures: &[f32],
     tick_rate: f64,
 ) -> Option<Vec<(f64, f64)>> {
     let curve = dyn_figure.curve()?;
@@ -567,7 +567,7 @@ pub fn eval_collider_slot(
     trace: &[Pose],
     traced: bool,
     capture_layout: Option<&CaptureLayout>,
-    captures: &[f64],
+    captures: &[f32],
     tick_rate: f64,
 ) -> ColliderData {
     match slot.repr() {
@@ -632,7 +632,7 @@ pub(super) fn capsule_chain_collider_data(
     trace: &[Pose],
     traced: bool,
     capture_layout: Option<&CaptureLayout>,
-    captures: &[f64],
+    captures: &[f32],
     tick_rate: f64,
 ) -> ColliderData {
     match dyn_figure.repr() {
